requirements.txt
discord.py==2.0.0  # Specify the version compatible with your code
requests==2.26.0  # Specify the version compatible with your code
python-dotenv==0.19.1  # Specify the version compatible with your code
